Move editor to module, add a test sandbox
