+++ clientapp/styles/pathway-editor.scss
@@ -1,23 +1,3 @@
-.cell {
-  border: 1px dashed #aaa;
-  padding: 0;
-  @include one-to-one-aspect;
-}
-
-.cell.drop {
-  background-color: rgba(255, 255, 0, 0.2);
-}
-
-.requirement {
-  text-align: center;
-  @include vertically-centered-children;
-
-  img {
-    max-width: 100%;
-    max-height: 100%;
-  }
-
-  &.core {
-    background-color: rgba(255, 0, 0, 0.2);
-  }
+.editor {
+  width: 100%;
 }
