+++ clientapp/templates/includes/editor.html
@@ -1,3 +1,1 @@
-<div>
-  <canvas></canvas>
-</div>
+<canvas class="editor" width="1000"></canvas>
