+++ package.json
@@ -41,6 +41,7 @@
     "mocha": "~1.15.1",
     "should": "~3.1.0",
     "precommit-hook": "~0.3.9",
-    "supertest": "~0.9.0"
+    "supertest": "~0.9.0",
+    "browserify": "~3.32.1"
   }
 }
